added confirm dialogs for all delete actions

+++ CourseManagementSystem.API/ClientApp/src/app/components/admin-list/admin-list.component.ts
@@ -6,6 +6,7 @@ import {WrapperVM} from '../../viewmodels/wrapperVM';
 import {CourseAdminService} from '../../services/course-admin.service';
 import {BsModalRef, BsModalService} from 'ngx-bootstrap/modal';
 import {ObservableWrapper} from '../../utils/observableWrapper';
+import {ConfirmDialogManager} from '../../utils/confirmDialogManager';
 
 /**
  * component representing list of course admins
@@ -34,6 +35,7 @@ export class AdminListComponent implements OnInit {
   private readonly courseAdminService: CourseAdminService;
   private bsModalRef: BsModalRef;
   private bsModalService: BsModalService;
+  private confirmDialogManager: ConfirmDialogManager;
   private observableWrapper: ObservableWrapper;
 
   constructor(courseService: CourseService, courseAdminService: CourseAdminService, bsModalService: BsModalService) {
@@ -41,6 +43,7 @@ export class AdminListComponent implements OnInit {
     this.courseAdminService = courseAdminService;
     this.bsModalService = bsModalService;
     this.observableWrapper = new ObservableWrapper(this.bsModalRef, this.bsModalService);
+    this.confirmDialogManager = new ConfirmDialogManager(this.bsModalRef, this.bsModalService);
   }
 
   ngOnInit() {
@@ -63,9 +66,14 @@ export class AdminListComponent implements OnInit {
    * @param adminId identifier of the admin to remove
    */
   public removeAdmin(adminId: string) {
-    this.courseAdminService.removeById(adminId).subscribe(() => {
-      this.reload();
-    });
+    this.confirmDialogManager.displayDialog(
+      'Remove an admin',
+      'Are you sure you want to remove this admin?',
+      () => {
+        this.courseAdminService.removeById(adminId).subscribe(() => {
+          this.reload();
+        });
+      });
   }
 
   /**
