+++ CourseManagementSystem.API/ClientApp/src/app/components/course-forum/course-forum.component.ts
@@ -6,6 +6,7 @@ import {RoleAuthService} from '../../services/role-auth.service';
 import {WrapperVM} from '../../viewmodels/wrapperVM';
 import {BsModalRef, BsModalService} from 'ngx-bootstrap/modal';
 import {ObservableWrapper} from '../../utils/observableWrapper';
+import {ConfirmDialogManager} from '../../utils/confirmDialogManager';
 
 @Component({
   selector: 'app-course-forum',
@@ -42,6 +43,7 @@ export class CourseForumComponent implements OnInit {
   private bsModalRef: BsModalRef;
   private bsModalService: BsModalService;
   private observableWrapper: ObservableWrapper;
+  private confirmDialogManager: ConfirmDialogManager;
 
   constructor(courseService: CourseService, forumPostService: ForumPostService, roleAuthService: RoleAuthService,
               bsModalService: BsModalService) {
@@ -49,6 +51,7 @@ export class CourseForumComponent implements OnInit {
     this.forumPostService = forumPostService;
     this.bsModalService = bsModalService;
     this.observableWrapper = new ObservableWrapper(this.bsModalRef, this.bsModalService);
+    this.confirmDialogManager = new ConfirmDialogManager(this.bsModalRef, this.bsModalService);
   }
 
   ngOnInit() {
@@ -60,9 +63,14 @@ export class CourseForumComponent implements OnInit {
    * @param post post to delete
    */
   public delete(post: ForumPostVM): void {
-    this.forumPostService.delete(post.id).subscribe(() => {
-      this.reloadPosts();
-    });
+    this.confirmDialogManager.displayDialog(
+      'Remove a forum post',
+      'Are you sure you want to Remove the selected post?',
+      () => {
+        this.forumPostService.delete(post.id).subscribe(() => {
+          this.reloadPosts();
+        });
+      });
   }
 
   /**
