+++ CourseManagementSystem.API/ClientApp/src/app/components/test-list/test-list.component.ts
@@ -5,6 +5,8 @@ import {RoleAuthService} from '../../services/role-auth.service';
 import {CourseTestService} from '../../services/course-test.service';
 import {DateTimeFormatter} from '../../utils/dateTimeFormatter';
 import {CourseTestUtils} from '../../utils/courseTestUtils';
+import {BsModalRef, BsModalService} from 'ngx-bootstrap/modal';
+import {ConfirmDialogManager} from '../../utils/confirmDialogManager';
 
 /**
  * component representing list of tests in a course
@@ -53,15 +55,21 @@ export class TestListComponent implements OnInit, OnChanges {
   private courseService: CourseService;
   private courseTestService: CourseTestService;
   private roleAuthService: RoleAuthService;
+  private bsModalRef: BsModalRef;
+  private bsModalService: BsModalService;
+  private confirmDialogManager: ConfirmDialogManager;
 
-  constructor(courseService: CourseService, roleAuthService: RoleAuthService, courseTestService: CourseTestService) {
+  constructor(courseService: CourseService, roleAuthService: RoleAuthService, courseTestService: CourseTestService,
+              bsModalService: BsModalService) {
     this.courseService = courseService;
     this.courseTestService = courseTestService;
     this.roleAuthService = roleAuthService;
+    this.bsModalService = bsModalService;
+    this.confirmDialogManager = new ConfirmDialogManager(this.bsModalRef, this.bsModalService);
   }
 
   ngOnInit() {
-      this.reloadTests();
+    this.reloadTests();
   }
 
   ngOnChanges(changes: SimpleChanges): void {
@@ -73,9 +81,14 @@ export class TestListComponent implements OnInit, OnChanges {
    * @param test test to delete
    */
   public deleteTest(test: CourseTestDetailsVM): void {
-    this.courseTestService.delete(test.id).subscribe(() => {
-      this.reloadTests();
-    });
+    this.confirmDialogManager.displayDialog(
+      'Delete an assignment',
+      'Are you sure you want to delete this assignment?',
+      () => {
+        this.courseTestService.delete(test.id).subscribe(() => {
+          this.reloadTests();
+        });
+      });
   }
 
   /**
